+++ BeatTogether.DedicatedServer.Messaging/Packets/PlayerLatencyPacket.cs
@@ -5,16 +5,16 @@ namespace BeatTogether.DedicatedServer.Messaging.Packets
 {
     public sealed class PlayerLatencyPacket : INetSerializable
     {
-        public float Latency { get; set; }
+        public long Latency { get; set; }
 
         public void ReadFrom(ref SpanBuffer reader)
         {
-            Latency = reader.ReadFloat32();
+            Latency = (long)reader.ReadUInt64();
         }
 
         public void WriteTo(ref SpanBuffer writer)
         {
-            writer.WriteFloat32(Latency);
+            writer.WriteUInt64((ulong)Latency);
         }
     }
 }
